requirements.txt
strawberry-graphql
fastapi
uvicorn
motor
python-dotenv
pydantic